+++ files/src/sdk/models/operations/patch_benefits_service_id_.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import serviceenablementresponse as shared_serviceenablementresponse
-from ..shared import serviceenablementupdate as shared_serviceenablementupdate
+from ...models.shared import serviceenablementresponse as shared_serviceenablementresponse
+from ...models.shared import serviceenablementupdate as shared_serviceenablementupdate
 from typing import Optional