+++ files/RELEASES.md
@@ -19,3 +19,14 @@ Based on:
 - [python v1.0.0] files
 ### Releases
 - [PyPI v1.0.0] https://pypi.org/project/wingspan_files/1.0.0 - files
+
+
+## 2023-11-13 01:24:29
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] files
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/wingspan_files/2.0.0 - files